requirements.txt
Django==3.2.16
django-admin-sortable2==1.0
Pillow=9.3.0